Restructure front-end app
