+++ app/views/index/index.html.erb
@@ -1,2 +1,2 @@
-<h1>Index</h1>
-<%= javascript_pack_tag 'hello_react' %>
+<%= javascript_pack_tag 'fab-look-frontend/index' %>
+<%= stylesheet_pack_tag 'fab-look-frontend/assets/stylesheets/app' %>
