+++ package.json
@@ -4,6 +4,10 @@
   "dependencies": {
     "@rails/webpacker": "3.5",
     "apollo-boost": "^0.1.27",
+    "apollo-cache-inmemory": "^1.4.3",
+    "apollo-link": "^1.2.8",
+    "apollo-link-batch-http": "^1.2.8",
+    "apollo-upload-client": "^10.0.0",
     "babel-preset-react": "^6.24.1",
     "graphql": "^14.1.1",
     "graphql-tag": "^2.10.1",
@@ -11,6 +15,8 @@
     "react": "^16.6.3",
     "react-apollo": "^2.4.1",
     "react-dom": "^16.6.3",
+    "react-dropzone": "^8.1.0",
+    "react-form": "^3.5.7",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
     "resolve-url-loader": "^3.0.0",
